// web-app/src/app/models/userDTO.interface.ts
export interface userDTO {
  usuId: number;
  usuName: string;
  usuLastName: string;
  usuTypeDni: string;
  usuNumDni: string;
  usuNumPhone: string;
  usuEmail: string;
  usuPassword: string;
  usuAttempts: number;
  usuProfilePicture: string;
  usuAdicionalArchive: string;
  rolId: number;
  estId: number;
  rolName: string;
  estadoName: string;
  estadoColor: string;
}